requirements.txt
opencv-python-headless>=4.10
numpy>=1.26
shapely>=2.0
fastdtw>=0.3.4
pykalman>=0.9.5
loguru>=0.7
scipy>=1.11
pydantic>=2.0
PyYAML>=6.0
